define scm status map in helper
