+++ lib/build.js
@@ -7,6 +7,7 @@ const deepcopy = require('deepcopy');
 const logger = require('screwdriver-logger');
 const { PR_JOB_NAME, EXTERNAL_TRIGGER } = require('screwdriver-data-schema').config.regex;
 const BaseModel = require('./base');
+const { SCM_STATE_MAP } = require('./helper');
 
 // Symbols for private members
 const executor = Symbol('executor');
@@ -15,22 +16,6 @@ const tokenGen = Symbol('tokenGen');
 const uiUri = Symbol('uiUri');
 const ABORT_CODE = 130; // 128 + SIGINT 2 (^C)
 const TEMPORAL_JWT_TIMEOUT = 12 * 60; // 12 hours in minutes
-
-// SCM handles only SUCCESS, PENDING & FAILURE status
-const SCM_STATE_MAP = {
-    ABORTED: 'FAILURE',
-    CREATED: 'PENDING',
-    FAILURE: 'FAILURE',
-    QUEUED: 'PENDING',
-    RUNNING: 'PENDING',
-    SUCCESS: 'SUCCESS',
-    BLOCKED: 'FAILURE',
-    UNSTABLE: 'FAILURE',
-    COLLAPSED: 'FAILURE',
-    FIXED: 'SUCCESS',
-    PENDING: 'PENDING',
-    FROZEN: 'FAILURE'
-};
 
 /**
  * Get the array of ids for jobs that match the names passed in
