+++ lib/helper.js
@@ -5,6 +5,21 @@ const hoek = require('@hapi/hoek');
 const dayjs = require('dayjs');
 const TEMPLATE_NAME_REGEX_WITH_NAMESPACE = schema.config.regex.FULL_TEMPLATE_NAME_WITH_NAMESPACE;
 const formatDate = dateTime => dayjs(dateTime).format('YYYY-MM-DD');
+// SCM handles only SUCCESS, PENDING & FAILURE status
+const SCM_STATE_MAP = {
+    ABORTED: 'FAILURE',
+    CREATED: 'PENDING',
+    FAILURE: 'FAILURE',
+    QUEUED: 'PENDING',
+    RUNNING: 'PENDING',
+    SUCCESS: 'SUCCESS',
+    BLOCKED: 'FAILURE',
+    UNSTABLE: 'FAILURE',
+    COLLAPSED: 'FAILURE',
+    FIXED: 'SUCCESS',
+    PENDING: 'PENDING',
+    FROZEN: 'FAILURE'
+};
 
 /**
  * Get the value of the annotation that matches name
@@ -228,5 +243,6 @@ module.exports = {
     parseTemplateConfigName,
     getAllRecords,
     getBuildClusterName,
-    getToken
+    getToken,
+    SCM_STATE_MAP
 };
